streamlit_app/requirements.txt
streamlit
newspaper3k
readability-lxml
beautifulsoup4
lxml_html_clean
